Extract diagnostics from the rule and group distinguishable predicates in namespace
