+++ packages/alfa-rules/src/sia-er62/rule.ts
@@ -1,26 +1,20 @@
-import { Rule, Diagnostic } from "@siteimprove/alfa-act";
+import { Rule } from "@siteimprove/alfa-act";
 import { Array } from "@siteimprove/alfa-array";
 import { Cache } from "@siteimprove/alfa-cache";
 import { Color } from "@siteimprove/alfa-css";
 import { Device } from "@siteimprove/alfa-device";
 import { Element, Node, Text } from "@siteimprove/alfa-dom";
 import { Iterable } from "@siteimprove/alfa-iterable";
-import { Equatable } from "@siteimprove/alfa-equatable";
-import { Hash, Hashable } from "@siteimprove/alfa-hash";
-import { Serializable } from "@siteimprove/alfa-json";
 import { Map } from "@siteimprove/alfa-map";
 import { Option, None } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Refinement } from "@siteimprove/alfa-refinement";
 import { Err, Ok, Result } from "@siteimprove/alfa-result";
 import { Context } from "@siteimprove/alfa-selector";
-import { Sequence } from "@siteimprove/alfa-sequence";
 import { Set } from "@siteimprove/alfa-set";
-import { Property, Style } from "@siteimprove/alfa-style";
+import { Style } from "@siteimprove/alfa-style";
 import { Criterion } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
-
-import * as json from "@siteimprove/alfa-json";
 
 import { Contrast } from "../../src/common/diagnostic/contrast";
 import { expectation } from "../common/expectation";
@@ -37,17 +31,16 @@ import {
 } from "../common/predicate";
 import { Stability } from "../tags/stability";
 
-import { Serialise } from "./serialise";
+import {
+  ComputedStyles,
+  DistinguishingStyles,
+  ExtendedDiagnostics,
+} from "./diagnostics";
 
 const { isElement } = Element;
 const { isText } = Text;
 const { or, not, test } = Predicate;
 const { and } = Refinement;
-
-export type ExtendedDiagnostics = {
-  computedStyles: ComputedStyles;
-  contrastPairings: Contrast.Pairing[];
-};
 
 /**
  * This version of R62 accepts differences in `font-family`, and differences
@@ -153,11 +146,17 @@ export default Rule.Atomic.of<Page, Element>({
             container,
             device,
             context
-          ).get();
-          const linkForegrounds = getForeground(link, device, context).get();
+          );
+          const containerColors = containerForegrounds.isOk()
+            ? containerForegrounds.get()
+            : [];
+          const linkForegrounds = getForeground(link, device, context);
+          const linkColors = linkForegrounds.isOk()
+            ? linkForegrounds.get()
+            : [];
           return [
-            ...Array.flatMap(containerForegrounds, (containerColor) =>
-              Array.map(linkForegrounds, (linkColor) => {
+            ...Array.flatMap(containerColors, (containerColor) =>
+              Array.map(linkColors, (linkColor) => {
                 return Contrast.Pairing.of(
                   containerColor,
                   linkColor,
@@ -177,9 +176,17 @@ export default Rule.Atomic.of<Page, Element>({
               // distinguishable from *all* non-link elements in order to be good.
               const hasDistinguishableStyle = nonLinkElements.some(
                 (container) =>
-                  isDistinguishable(container, device, context)(link) ||
+                  Distinguishable.isDistinguishable(
+                    container,
+                    device,
+                    context
+                  )(link) ||
                   (context.isHovered(target) &&
-                    hasDistinguishableCursor(container, device, context)(link))
+                    Distinguishable.hasDistinguishableCursor(
+                      container,
+                      device,
+                      context
+                    )(link))
               );
               const distinguishableContrast = Set.from(
                 Array.from(nonLinkElements)
@@ -331,371 +338,206 @@ function hasNonLinkText(device: Device): Predicate<Element> {
   };
 }
 
-function isDistinguishable(
-  container: Element,
-  device: Device,
-  context: Context = Context.empty()
-): Predicate<Element> {
-  return or(
-    // Things like text decoration and backgrounds risk blending with the
-    // container element. We therefore need to check if these can be distinguished
-    // from what the container element might itself set.
-    hasDistinguishableBackground(container, device, context),
-    hasDistinguishableContrast(container, device, context),
-    hasDistinguishableFont(container, device, context),
-    hasDistinguishableTextDecoration(container, device, context),
-    hasDistinguishableVerticalAlign(container, device, context),
-    // We consider the mere presence of borders or outlines on the element as
-    // distinguishable features. There's of course a risk of these blending with
-    // other features of the container element, such as its background, but this
-    // should hopefully not happen (too often) in practice. When it does, we
-    // risk false negatives.
-    hasBorder(device, context),
-    hasOutline(device, context)
-  );
-}
-
-function hasDistinguishableTextDecoration(
-  container: Element,
-  device: Device,
-  context?: Context
-): Predicate<Element> {
-  return (element) =>
-    test(not(hasTextDecoration(device, context)), container) &&
-    test(hasTextDecoration(device, context), element);
-}
-
-/**
- * Check if an element has a distinguishable background from the given container
- * element.
- *
- * @remarks
- * This predicate currently only considers `background-color` and
- * `background-image` as a possibly distinguishable background. Other
- * `background-*` properties should ideally also be considered.
- *
- * Additionally, this predicate do not handle transparency in the topmost layer.
- * The exact same (partly transparent) `background-color` or `background-image`
- * could be on top of a different (opaque) background and thus creates a
- * difference. However, in these cases the (lower layer) distinguishing
- * background would be on an ancestor of the link but of no non-link text (in
- * order to be distinguishing), so should be found when looking at the ancestors
- * of the link.
- *
- * Lastly, this does not account for absolutely positioned backgrounds from
- * random nodes in the DOM. Using these to push an image below links in
- * paragraph sounds so crazy (from a sheer code maintenance point of view) that
- * this hopefully won't be a problem.
- */
-function hasDistinguishableBackground(
-  container: Element,
-  device: Device,
-  context?: Context
-): Predicate<Element> {
-  const colorReference = Style.from(container, device, context).computed(
-    "background-color"
-  ).value;
-
-  const imageReference = Style.from(container, device, context).computed(
-    "background-image"
-  ).value;
-
-  return or(
-    hasComputedStyle(
-      "background-color",
-      not(
-        // If the background is fully transparent, we assume it will end up
-        // being the same as the container. Intermediate backgrounds may change
-        // that, but these would need to be set on ancestor of the link and of
-        // no non-link text, so will be caught in one of the other comparisons.
-        (color) => Color.isTransparent(color) || color.equals(colorReference)
-      ),
-      device,
-      context
-    ),
-    // Any difference in `background-image` is considered enough. If different
-    // `background-image` ultimately yield the same background (e.g. the same
-    // image at two different URLs), this creates false negatives.
-    hasComputedStyle(
-      "background-image",
-      not((image) => image.equals(imageReference)),
-      device,
-      context
-    )
-  );
-}
-
-function hasDistinguishableContrast(
-  container: Element,
-  device: Device,
-  context?: Context
-): Predicate<Element> {
-  return (link) =>
-    getForeground(container, device, context).some((containerColors) =>
-      getForeground(link, device, context).some((linkColors) => {
-        const contrastValues = [
-          ...Array.flatMap(containerColors, (containerColor) =>
-            Array.map(linkColors, (linkColor) => {
-              return contrast(containerColor, linkColor);
-            })
-          ),
-        ];
-
-        for (const contrastValue of contrastValues) {
-          // If at least one of the contrast values are bigger than the threshold, the link is marked distinguisable
-          if (contrastValue >= 3) {
-            return true;
-          }
-        }
-        return false;
-      })
-    );
-}
-
-/**
- * Check if an element has a different font weight or family than its container.
- *
- * This is brittle and imperfect but removes a strong pain point until we find
- * a better solution.
- */
-
-function hasDistinguishableFont(
-  container: Element,
-  device: Device,
-  context?: Context
-): Predicate<Element> {
-  const style = Style.from(container, device, context);
-
-  const referenceWeight = style.computed("font-weight").value;
-  const referenceFamily = Option.from(
-    style.computed("font-family").value.values[0]
-  );
-
-  return or(
-    hasComputedStyle(
-      "font-weight",
-      not((weight) => weight.equals(referenceWeight)),
-      device,
-      context
-    ),
-    hasComputedStyle(
-      "font-family",
-      not((family) => Option.from(family.values[0]).equals(referenceFamily)),
-      device,
-      context
-    )
-  );
-}
-
-function hasDistinguishableVerticalAlign(
-  container: Element,
-  device: Device,
-  context?: Context
-): Predicate<Element> {
-  const reference = Style.from(container, device, context).computed(
-    "vertical-align"
-  ).value;
-
-  return hasComputedStyle(
-    "vertical-align",
-    not((alignment) => alignment.equals(reference)),
-    device,
-    context
-  );
-}
-
-function hasDistinguishableCursor(
-  container: Element,
-  device: Device,
-  context?: Context
-): Predicate<Element> {
-  // Checking if there is a custom cursor, otherwise grabbing the built-in
-  function getFirstCursor(style: Style.Computed<"cursor">) {
-    return style.values[0].values.length !== 0
-      ? style.values[0].values[0]
-      : style.values[1];
-  }
-  const containerCursorStyle = Style.from(container, device, context).computed(
-    "cursor"
-  ).value;
-
-  // We assume that the first custom cursor, if any, will never fail to load
-  // and thus don't try to default further.
-  const reference = getFirstCursor(containerCursorStyle);
-
-  return hasComputedStyle(
-    "cursor",
-    not((cursor) => getFirstCursor(cursor).equals(reference)),
-    device,
-    context
-  );
-}
-
-type Name = Property.Name | Property.Shorthand.Name;
-
-export class ComputedStyles implements Equatable, Hashable, Serializable {
-  public static of(
-    style: Iterable<readonly [Name, string]> = []
-  ): ComputedStyles {
-    return new ComputedStyles(Map.from(style));
-  }
-
-  private readonly _style: Map<Name, string>;
-
-  private constructor(style: Map<Name, string>) {
-    this._style = style;
-  }
-
-  public get style(): Map<Name, string> {
-    return this._style;
-  }
-
-  public with(
-    ...styles: ReadonlyArray<readonly [Name, string]>
-  ): ComputedStyles {
-    return ComputedStyles.of([...this._style, ...styles]);
-  }
-
-  public equals(value: ComputedStyles): boolean;
-
-  public equals(value: unknown): value is this;
-
-  public equals(value: unknown): boolean {
-    return value instanceof ComputedStyles && value._style.equals(this._style);
-  }
-
-  public hash(hash: Hash): void {
-    this._style.hash(hash);
-  }
-
-  public toJSON(): ComputedStyles.JSON {
-    return {
-      style: this._style.toJSON(),
-    };
-  }
-}
-
-export namespace ComputedStyles {
-  export interface JSON {
-    [key: string]: json.JSON;
-    style: Map.JSON<Name, string>;
-  }
-
-  export function isComputedStyles(value: unknown): value is ComputedStyles {
-    return value instanceof ComputedStyles;
-  }
-
-  export function from(
-    element: Element,
+namespace Distinguishable {
+  export function isDistinguishable(
+    container: Element,
     device: Device,
-    target: Element,
     context: Context = Context.empty()
-  ): ComputedStyles {
-    const style = Style.from(element, device, context);
-
-    const border = (["color", "style", "width"] as const).map((property) =>
-      Serialise.borderShorthand(style, property)
-    );
-
-    const cursor = context.isHovered(target)
-      ? [["cursor", Serialise.getLonghand(style, "cursor")] as const]
-      : [];
-
-    return ComputedStyles.of(
-      [
-        ...border,
-        ...cursor,
-        ["color", Serialise.getLonghand(style, "color")] as const,
-        ["font", Serialise.font(style)] as const,
-        [
-          "vertical-align",
-          Serialise.getLonghand(style, "vertical-align"),
-        ] as const,
-        ["background", Serialise.background(style)] as const,
-        ["outline", Serialise.outline(style)] as const,
-        ["text-decoration", Serialise.textDecoration(style)] as const,
-        // ["cursor", Serialise.getLonghand(style, "cursor")] as const,
-      ].filter(([_, value]) => value !== "")
-    );
-  }
-}
-
-export class DistinguishingStyles extends Diagnostic {
-  public static of(
-    message: string,
-    defaultStyles: Iterable<Result<ExtendedDiagnostics>> = Sequence.empty(),
-    hoverStyles: Iterable<Result<ExtendedDiagnostics>> = Sequence.empty(),
-    focusStyles: Iterable<Result<ExtendedDiagnostics>> = Sequence.empty()
-  ): DistinguishingStyles {
-    return new DistinguishingStyles(
-      message,
-      Sequence.from(defaultStyles),
-      Sequence.from(hoverStyles),
-      Sequence.from(focusStyles)
+  ): Predicate<Element> {
+    return or(
+      // Things like text decoration and backgrounds risk blending with the
+      // container element. We therefore need to check if these can be distinguished
+      // from what the container element might itself set.
+      hasDistinguishableBackground(container, device, context),
+      hasDistinguishableContrast(container, device, context),
+      hasDistinguishableFont(container, device, context),
+      hasDistinguishableTextDecoration(container, device, context),
+      hasDistinguishableVerticalAlign(container, device, context),
+      // We consider the mere presence of borders or outlines on the element as
+      // distinguishable features. There's of course a risk of these blending with
+      // other features of the container element, such as its background, but this
+      // should hopefully not happen (too often) in practice. When it does, we
+      // risk false negatives.
+      hasBorder(device, context),
+      hasOutline(device, context)
     );
   }
 
-  private readonly _defaultStyles: Sequence<Result<ExtendedDiagnostics>>;
-  private readonly _hoverStyles: Sequence<Result<ExtendedDiagnostics>>;
-  private readonly _focusStyles: Sequence<Result<ExtendedDiagnostics>>;
-
-  private constructor(
-    message: string,
-    defaultStyles: Sequence<Result<ExtendedDiagnostics>>,
-    hoverStyles: Sequence<Result<ExtendedDiagnostics>>,
-    focusStyles: Sequence<Result<ExtendedDiagnostics>>
-  ) {
-    super(message);
-    this._defaultStyles = defaultStyles;
-    this._hoverStyles = hoverStyles;
-    this._focusStyles = focusStyles;
+  function hasDistinguishableTextDecoration(
+    container: Element,
+    device: Device,
+    context?: Context
+  ): Predicate<Element> {
+    return (element) =>
+      test(not(hasTextDecoration(device, context)), container) &&
+      test(hasTextDecoration(device, context), element);
   }
 
-  public get defaultStyles(): Iterable<Result<ExtendedDiagnostics>> {
-    return this._defaultStyles;
-  }
+  /**
+   * Check if an element has a distinguishable background from the given container
+   * element.
+   *
+   * @remarks
+   * This predicate currently only considers `background-color` and
+   * `background-image` as a possibly distinguishable background. Other
+   * `background-*` properties should ideally also be considered.
+   *
+   * Additionally, this predicate do not handle transparency in the topmost layer.
+   * The exact same (partly transparent) `background-color` or `background-image`
+   * could be on top of a different (opaque) background and thus creates a
+   * difference. However, in these cases the (lower layer) distinguishing
+   * background would be on an ancestor of the link but of no non-link text (in
+   * order to be distinguishing), so should be found when looking at the ancestors
+   * of the link.
+   *
+   * Lastly, this does not account for absolutely positioned backgrounds from
+   * random nodes in the DOM. Using these to push an image below links in
+   * paragraph sounds so crazy (from a sheer code maintenance point of view) that
+   * this hopefully won't be a problem.
+   */
+  function hasDistinguishableBackground(
+    container: Element,
+    device: Device,
+    context?: Context
+  ): Predicate<Element> {
+    const colorReference = Style.from(container, device, context).computed(
+      "background-color"
+    ).value;
 
-  public get hoverStyles(): Iterable<Result<ExtendedDiagnostics>> {
-    return this._hoverStyles;
-  }
+    const imageReference = Style.from(container, device, context).computed(
+      "background-image"
+    ).value;
 
-  public get focusStyles(): Iterable<Result<ExtendedDiagnostics>> {
-    return this._focusStyles;
-  }
-
-  public equals(value: DistinguishingStyles): boolean;
-
-  public equals(value: unknown): value is this;
-
-  public equals(value: unknown): boolean {
-    return (
-      value instanceof DistinguishingStyles &&
-      value._defaultStyles.equals(this._defaultStyles) &&
-      value._hoverStyles.equals(this._hoverStyles) &&
-      value._focusStyles.equals(this._focusStyles)
+    return or(
+      hasComputedStyle(
+        "background-color",
+        not(
+          // If the background is fully transparent, we assume it will end up
+          // being the same as the container. Intermediate backgrounds may change
+          // that, but these would need to be set on ancestor of the link and of
+          // no non-link text, so will be caught in one of the other comparisons.
+          (color) => Color.isTransparent(color) || color.equals(colorReference)
+        ),
+        device,
+        context
+      ),
+      // Any difference in `background-image` is considered enough. If different
+      // `background-image` ultimately yield the same background (e.g. the same
+      // image at two different URLs), this creates false negatives.
+      hasComputedStyle(
+        "background-image",
+        not((image) => image.equals(imageReference)),
+        device,
+        context
+      )
     );
   }
 
-  public toJSON(): DistinguishingStyles.JSON {
-    return {
-      ...super.toJSON(),
-      defaultStyle: this._defaultStyles.toJSON(),
-      hoverStyle: this._hoverStyles.toJSON(),
-      focusStyle: this._focusStyles.toJSON(),
-    };
-  }
-}
+  function hasDistinguishableContrast(
+    container: Element,
+    device: Device,
+    context?: Context
+  ): Predicate<Element> {
+    return (link) =>
+      getForeground(container, device, context).some((containerColors) =>
+        getForeground(link, device, context).some((linkColors) => {
+          const contrastValues = [
+            ...Array.flatMap(containerColors, (containerColor) =>
+              Array.map(linkColors, (linkColor) => {
+                return contrast(containerColor, linkColor);
+              })
+            ),
+          ];
 
-export namespace DistinguishingStyles {
-  export interface JSON extends Diagnostic.JSON {
-    defaultStyle: Sequence.JSON<Result<ExtendedDiagnostics>>;
-    hoverStyle: Sequence.JSON<Result<ExtendedDiagnostics>>;
-    focusStyle: Sequence.JSON<Result<ExtendedDiagnostics>>;
+          for (const contrastValue of contrastValues) {
+            // If at least one of the contrast values are bigger than the threshold, the link is marked distinguisable
+            if (contrastValue >= 3) {
+              return true;
+            }
+          }
+          return false;
+        })
+      );
   }
 
-  export function isDistinguishingStyles(
-    value: unknown
-  ): value is DistinguishingStyles {
-    return value instanceof DistinguishingStyles;
+  /**
+   * Check if an element has a different font weight or family than its container.
+   *
+   * This is brittle and imperfect but removes a strong pain point until we find
+   * a better solution.
+   */
+
+  function hasDistinguishableFont(
+    container: Element,
+    device: Device,
+    context?: Context
+  ): Predicate<Element> {
+    const style = Style.from(container, device, context);
+
+    const referenceWeight = style.computed("font-weight").value;
+    const referenceFamily = Option.from(
+      style.computed("font-family").value.values[0]
+    );
+
+    return or(
+      hasComputedStyle(
+        "font-weight",
+        not((weight) => weight.equals(referenceWeight)),
+        device,
+        context
+      ),
+      hasComputedStyle(
+        "font-family",
+        not((family) => Option.from(family.values[0]).equals(referenceFamily)),
+        device,
+        context
+      )
+    );
+  }
+
+  function hasDistinguishableVerticalAlign(
+    container: Element,
+    device: Device,
+    context?: Context
+  ): Predicate<Element> {
+    const reference = Style.from(container, device, context).computed(
+      "vertical-align"
+    ).value;
+
+    return hasComputedStyle(
+      "vertical-align",
+      not((alignment) => alignment.equals(reference)),
+      device,
+      context
+    );
+  }
+
+  export function hasDistinguishableCursor(
+    container: Element,
+    device: Device,
+    context?: Context
+  ): Predicate<Element> {
+    // Checking if there is a custom cursor, otherwise grabbing the built-in
+    function getFirstCursor(style: Style.Computed<"cursor">) {
+      return style.values[0].values.length !== 0
+        ? style.values[0].values[0]
+        : style.values[1];
+    }
+    const containerCursorStyle = Style.from(
+      container,
+      device,
+      context
+    ).computed("cursor").value;
+
+    // We assume that the first custom cursor, if any, will never fail to load
+    // and thus don't try to default further.
+    const reference = getFirstCursor(containerCursorStyle);
+
+    return hasComputedStyle(
+      "cursor",
+      not((cursor) => getFirstCursor(cursor).equals(reference)),
+      device,
+      context
+    );
   }
 }
