+++ packages/alfa-rules/tsconfig.json
@@ -83,6 +83,7 @@
     "src/tags/stability.ts",
     "src/tags/scope.ts",
     "src/sia-er62/rule.ts",
+    "src/sia-er62/diagnostics.ts",
     "src/sia-er62/serialise.ts",
     "src/sia-er87/rule.ts",
     "src/sia-r1/rule.ts",
